refactor: Filestructure
Moved around files within GenerateReceivingModal for future work on API layer
